requirements.txt
kafka-python==2.2.15
redis==4.5.1
flask==2.3.2
flask-socketio==5.5.1